move unit tests under packages under io.personium.core.*
