+++ src/test/java/io/personium/core/rs/odata/EscapeResponseBodyTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.personium.test.unit.core.rs.odata;
+package io.personium.core.rs.odata;
 
 import static org.junit.Assert.assertEquals;
