+++ src/test/java/io/personium/core/rs/odata/SentMessageValidateTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.personium.test.unit.core.rs.odata.validate;
+package io.personium.core.rs.odata;
 
 import static org.junit.Assert.assertEquals;
