+++ src/test/java/io/personium/test/RegressionTests.java
@@ -40,6 +40,7 @@ import junit.framework.TestSuite;
         io.personium.core.model.impl.es.accessor.AllTests.class,
         io.personium.core.model.lock.AllTests.class,
         io.personium.core.rs.AllTests.class,
+        io.personium.core.rs.odata.AllTests.class,
         io.personium.test.jersey.AllTests.class,
         io.personium.test.jersey.bar.AllTests.class,
         io.personium.test.jersey.box.AllTests.class,
@@ -56,7 +57,6 @@ import junit.framework.TestSuite;
         io.personium.test.jersey.concurrent.AllTests.class,
         io.personium.test.performance.box.odatacol.AllTests.class,
         io.personium.test.unit.core.AllTests.class,
-        io.personium.test.unit.core.rs.odata.AllTests.class,
         io.personium.test.unit.core.rs.odata.validate.AllTests.class
 })
 @IncludeCategory(Regression.class)
