+++ src/test/java/io/personium/test/UnitTests.java
@@ -34,8 +34,8 @@ import junit.framework.TestSuite;
         io.personium.core.auth.AllTests.class,
         io.personium.core.bar.AllTests.class,
         io.personium.core.odata.AllTests.class,
-        io.personium.core.rs.odata.AllTests.class,
         io.personium.core.rs.AllTests.class,
+        io.personium.core.rs.odata.AllTests.class,
         io.personium.core.utils.AllTests.class,
         io.personium.core.model.file.AllTests.class,
         io.personium.core.model.impl.es.AllTests.class,
@@ -58,9 +58,7 @@ import junit.framework.TestSuite;
         io.personium.test.jersey.cell.ctl.AllTests.class,
         io.personium.test.jersey.concurrent.AllTests.class,
         io.personium.test.performance.box.odatacol.AllTests.class,
-        io.personium.test.unit.core.AllTests.class,
-        io.personium.test.unit.core.rs.odata.AllTests.class,
-        io.personium.test.unit.core.rs.odata.validate.AllTests.class
+        io.personium.test.unit.core.AllTests.class
 })
 @IncludeCategory(Integration.class)
 public class UnitTests extends TestSuite {
